fix : haveDesign <-> noDesign 케이스 swap
